// angularProject/src/app/api/model/order.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderStatus } from './orderStatus';
import { PaymentStatus } from './paymentStatus';


export interface Order { 
    id?: number;
    keyUsage?: string | null;
    amountPaid?: number | null;
    symbol?: string | null;
    customerId?: number | null;
    creditCount?: number | null;
    orderGuid?: string;
    storeId?: number;
    billingAddressId?: number;
    orderStatusId?: number;
    paymentStatusId?: number;
    paymentMethodSystemName?: string | null;
    customerCurrencyCode?: string | null;
    currencyRate?: number;
    orderSubTotalDiscountInclTax?: number;
    orderSubTotalDiscountExclTax?: number;
    orderDiscount?: number;
    orderTotal?: number;
    paidDateUtc?: string | null;
    createdOnUtc?: string;
    customOrderNumber?: string | null;
    redeemedCoinsEntryId?: number | null;
    productType?: string | null;
    paymentType?: string | null;
    productAttributeId?: number;
    advisorId?: number;
    orderTax?: number;
    orderStatus?: OrderStatus;
    paymentStatus?: PaymentStatus;
}
export namespace Order {
}